refactor(library): retire QuickActionLaunchModal shim; migrate tests to ModelLaunchModal
QuickActionLaunchModal had no remaining production consumers; launchWithModelPicker
uses ModelLaunchModal directly. Migrate the 5 cases the old test exclusively
covered (lists-only-enabled-providers, switching-provider-resets-model,
fallback-notice-shown/hidden, Cancel-does-not-fire-onConfirm) into
modelLaunchModal.test.ts using a two-provider enabledProviders fixture.
Delete the now-empty shim source and its test file.
